DDF-1772 Remove references to org.apache.servicemix.bundles.lucene in offline-gazetteer and refactored spatial geocoding packages
