+++ catalog/spatial/geocoding/pom.xml
@@ -29,9 +29,8 @@
         <module>spatial-geocoding-create</module>
         <module>spatial-geocoding-extract</module>
         <module>spatial-geocoding-geocoder</module>
-        <module>spatial-geocoding-index</module>
+        <module>spatial-geocoding-lucene</module>
         <module>spatial-geocoding-localsearch</module>
-        <module>spatial-geocoding-query</module>
         <module>spatial-geocoding-websearch</module>
         <module>spatial-geocoding-admin-module</module>
     </modules>
