+++ catalog/spatial/geocoding/spatial-geocoding-lucene/src/main/resources/OSGI-INF/metatype/metatype.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+/**
+ * Copyright (c) Codice Foundation
+ *
+ * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
+ * <http://www.gnu.org/licenses/lgpl.html>.
+ *
+ **/
+
+ -->
+<metatype:MetaData xmlns:metatype="http://www.osgi.org/xmlns/metatype/v1.0.0">
+
+    <OCD name="GeoNames Indexer" id="org.codice.ddf.spatial.geocoding.index.properties">
+        <AD description="Specifies the directory to hold the GeoNames index."
+            name="Index Location" id="indexLocation" required="true" type="String"
+            default="data/geonames-index"
+        />
+    </OCD>
+
+    <Designate pid="org.codice.ddf.spatial.geocoding.index.properties">
+        <Object ocdref="org.codice.ddf.spatial.geocoding.index.properties"/>
+    </Designate>
+
+</metatype:MetaData>
