+++ catalog/spatial/geocoding/spatial-geocoding-lucene/src/test/java/org/codice/ddf/spatial/geocoding/index/TestGeoNamesLuceneIndexer.java
